Remove sequence export from Logger component
